tweak the CSS
Make the Image div "crushable" meaning it will resize and the left div
won't get pushed off the screen.

Add in some 1ex padding and margins between important blocks of content.
Specifically the image div and the content body, also the menu and the
copy text.

Reset the top margin on Headings so that everything lines up properly.